No.231 solved.
if a number is power of two.

it must be greater than 0,

and its binary number must have one 1 in highest position other all are 0;

it means this number minus 1, its binary number all are 1;

let make them do & operation,

if result is 0,this number is power of two,
if result is not 0,this number isnot power of two.

for example, 4's binary number is 0100, 3's binary number is 0011,
0100
&
0011
0000, the result is 0, so 4 is power of two,

8's binary number is 1000, 7's binary number is 0111,
1000
&
0111
0000, the result is 0, so 4 is power of two,

and counter-example , 9's binary number is 1001, 8's binary number is 1000,
1001
&
1000
1000,the result is 8, so 9 isnot power of two.

so the code is just one line: return n > 0 && (n&(n-1)) == 0;