No 237 solved.
in fact ,i dont understanded this question very well when i met it first time,

i konw tree,
i know search,
but sorry i dont konw search tree,

and when i get what search tree means,i realized quickly how to solved .

search tree means:
the root tree value is the middle value of the whole tree,
so it minus it's left side value the result is bigger than 0;
and it minus it's right side value the result is less than 0;

if tree p and tree q on the different side of tree root ,
tree root is result

we know this characteristic of search tree ,
so we can create a recursion method  .

we can let root's value minus value of other trees value and let them multiplication ,
if the result less than 0, we can make sure that the root is result.
if the result bigger than 0, we know they are on same side.
we also need get  the side they are on.
if they  both bigger than value of root tree,  we continu do the recursion  with right side of root tree and tree p,q,
and if they both less than value of root tree, we continu do the recursion  with left side of root tree and tree p,q until get result.