No.242 solved
this question want us judge if chars of two String  are same and its appear frequency is all same.

i solved this by three ways.

first way: map.  record every char of one string as KEY of this map,
 and it's appear frequency as VAlUE of this map. and then traversal
chars of another string,if map doesn't contains this char, certainly the result is false. if map contains this char,get the value of this key,and minus 1.
when we traversal another String finish ,go on traversal the map, check if every value of key is 0, if not ,the result is false, if it is, the result is true.

it is too stupid. but it can used by every Encoding.

then i tried second way:
get char array of two string, and sorted it, return they are same or not.
question solved.. so easy. but traversal too many times.

so i tried the third way:
use a new array named 'COUNT' ,it's length is 26. it's type is Integer.
traversal two string  to get two chars,
make it minus 'a' to get their really index,
one of them add 1 ,another minus 1,
put the result in 'COUNT'

for example : String A = "anagram", B = "nagaram";
String A first char is 'a',   String B first char is 'n'
thay all minus 'a' ,get their really index,
A's  really index is 0, B's really index is 13,
now make index 0 of COUNT minus 1, index 13 of COUNT add 1,
the COUNT is
 [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];

go on,
String A second char is 'n', String B second char is 'a',
thay all minus 'a' ,get their really index,
A's really index is 13, B's really index is 0,
now make index 13 of COUNT minus 1, index 0 of COUNT add 1,
the COUNT is
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
do it again until end of string.

if this two string have same word and it's appear frequency is also same,all element of 'COUNT' is 0. .
we just need traversal 'COUNT',
if all element is 0,the result is true,or not  the result is false.