No.258 solved
Brute Force method needless to explane too much.

Math method is very interesting:

origin number is n= 100a +10b+c;
result number is addn = a+b+c;
so they disparity is n - addn = 99a+9b = 9(9a+b);
it means every time from origin number to temp number,
D-value is times of 9!
so we make origin number modularization of '9' many times to get result.

if this numer modularization of '9' is zero, the result is '9'
we can also get this regular from that.

and don't forget the border '9'!

all of that are already in my code.